id long

+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -26,29 +26,29 @@ public class BookingController {
     private final BookingService bookingService;
 
     @PostMapping
-    public BookingResponseDto create(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) int userId,
+    public BookingResponseDto create(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) Long userId,
                                      @Valid @RequestBody BookingDto booking) {
         log.info("Начало выполнения по эндпоинту @PostMapping create");
         return bookingService.create(userId, booking);
     }
 
     @PatchMapping("/{bookingId}")
-    public BookingResponseDto update(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) int userId,
-                                     @PathVariable int bookingId,
+    public BookingResponseDto update(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) Long userId,
+                                     @PathVariable Long bookingId,
                                      @RequestParam boolean approved) {
         log.info("Начало выполнения по эндпоинту @PatchMapping update");
         return bookingService.update(userId, bookingId, approved);
     }
 
     @GetMapping("/{bookingId}")
-    public BookingResponseDto getBooking(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) int userId,
-                                         @PathVariable int bookingId) {
+    public BookingResponseDto getBooking(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) Long userId,
+                                         @PathVariable Long bookingId) {
         log.info("Начало выполнения по эндпоинту @GetMapping getBooking");
         return bookingService.getBooking(userId, bookingId);
     }
 
     @GetMapping
-    public List<BookingResponseDto> getBookings(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) int userId,
+    public List<BookingResponseDto> getBookings(@RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) Long userId,
                                                 @RequestParam(required = false, defaultValue = "ALL") String state) {
         log.info("Начало выполнения по эндпоинту @GetMapping {bookingId}");
         return bookingService.getBookings(userId, state);
@@ -57,7 +57,7 @@ public class BookingController {
 
     @GetMapping("/owner")
     public List<BookingResponseDto> getBookingsOwner(
-            @RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) int userId,
+            @RequestHeader(HttpHeadersConstants.X_SHARER_USER_ID) Long userId,
             @RequestParam(required = false, defaultValue = "ALL") String state) {
         log.info("Начало выполнения по эндпоинту @GetMapping getBookingsOwner");
         return bookingService.getBookingsOwner(userId, state);
