+++ src/main/java/ru/practicum/shareit/booking/BookingRepository.java
@@ -10,24 +10,24 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
+public interface BookingRepository extends JpaRepository<Booking, Long> {
 
-    Optional<List<Booking>> findByItemOwner_Id(int ownerId);
+    Optional<List<Booking>> findByItemOwner_Id(Long ownerId);
 
-    Optional<Booking> findByBookerId(int bookerId);
+    Optional<Booking> findByBookerId(Long bookerId);
 
-    List<Booking> findAllByBookerId(int bookerId, Sort sort);
+    List<Booking> findAllByBookerId(Long bookerId, Sort sort);
 
-    List<Booking> findByBookerIdAndEndIsBefore(int bookerId, LocalDateTime end, Sort sort);
+    List<Booking> findByBookerIdAndEndIsBefore(Long bookerId, LocalDateTime end, Sort sort);
 
     @Query("SELECT b FROM Booking b WHERE b.booker.id = ?1 " +
             " AND b.start < ?2 AND b.end > ?2 " +
             " AND b.status = 'APPROVED' ORDER BY ?3")
-    List<Booking> findRelevantBookingsByBookingId(int bookerId, LocalDateTime currentTime, Sort sort);
+    List<Booking> findRelevantBookingsByBookingId(Long bookerId, LocalDateTime currentTime, Sort sort);
 
-    List<Booking> findByBookerIdAndEndAfter(int bookerId, LocalDateTime currentTime, Sort sort);
+    List<Booking> findByBookerIdAndEndAfter(Long bookerId, LocalDateTime currentTime, Sort sort);
 
-    List<Booking> findBookingByIdAndStatus(int bookerId, BookingStatus status, Sort sort);
+    List<Booking> findBookingByIdAndStatus(Long bookerId, BookingStatus status, Sort sort);
 
     List<Booking> findAllByItemIn(List<Item> items, Sort sort);
