+++ src/main/java/ru/practicum/shareit/item/CommentRepository.java
@@ -5,8 +5,8 @@ import ru.practicum.shareit.item.model.Comment;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    List<Comment> findByItemId(int itemId);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    List<Comment> findByItemId(Long itemId);
 
-    List<Comment> findByItemIdIn(List<Integer> itemIds);
+    List<Comment> findByItemIdIn(List<Long> itemIds);
 }
