+++ src/main/java/ru/practicum/shareit/item/ItemRepository.java
@@ -6,8 +6,8 @@ import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
-    List<Item> findItemsByOwnerId(int ownerId);
+public interface ItemRepository extends JpaRepository<Item, Long> {
+    List<Item> findItemsByOwnerId(Long ownerId);
 
     @Query("SELECT i FROM Item i " +
             "WHERE lower(i.name) like lower(concat('%', :text, '%')) " +
